arm: Make sure to at least run a certain number of rounds in 'log' functions
The value 30 is a blind guesstimate of what's sensible; with that, I
get log() returning values close to the real values (differing only in
the third digit or so).

I'm not sure exactly how the original logic for the number of rounds
is supposed to work; with any value over 16, it will previously
run 0 rounds and return 0.

Signed-off-by: Martin Storsjö <martin@martin.st>
(cherry picked from commit 41de4baaccba1364595b96d4339c157bc04234c9)