Make ARM softmath functions experimental, disabled by default
By default, just forward calls to msvcrt.dll for the functions
implemented with softmath routines.

The msvcrt math functions might not be fully C99 compliant, but
the current softfloat implementations are lacking in many other ways
(e.g. crashing due to infinite recursion for some input values, only
working for certain ranges of input values, etc.).

For functions that are missing in msvcrt.dll (log2*, exp2*, scalbn*),
provide small wrappers that fall back on log and pow.

This makes the libav testsuite succeed (when run in wine), when
built with clang for ARM with mingw-w64 - previously a number of
tests failed due to faulty math functions.

Signed-off-by: Martin Storsjö <martin@martin.st>