Change check for _WIN64 to check for __x86_64__ in assembler files
Lots of assembler files in the math subdir check for _WIN64 when they
actually mean to check for the x86_64 target CPU, rather than the target
OS.  This patch fixes it.  This helps Cygwin in the first place because
__x86_64__ is defined in both toolchains, while _WIN64 is only defined
on Mingw-w64 for x86_64.

Signed-off-by: Corinna Vinschen <vinschen@redhat.com>
(cherry picked from commit 40f03bb9ebc8c43235871329466b1b8a5794a27b)
Signed-off-by: Jonathan Yong <10walls@gmail.com>