Update definitions in 64bit ntdll.def to match the 32bit one.
This marks the math functions cos, fabs, atan and floor as DATA in
ntdll to mirror the behavior of the 32bit definitions.

This makes the behavior of the two consistent in that it forces to
have mingwex before ntdll in order to use these functions so you
can get the mingwex provided one.

It also makes it so you don't unknownly get the ntdll ones which may
have a different precision than what you'd expect.

Signed-off-by: Tamar Christina <tamar@zhox.com>