crt: More strictly flag functions in the unified msvcrt.def.in
Now the output from the unified msvcrt.def.in should be almost identical
to what it was before, except for some new *_dbg and __p_* functions.

The functions that were included for 32 and 64 bit x86 in the unification
were originally believed to be harmless, but some of them turned out to
be function that we have compatibility fallbacks for, that were missed
when libmsvcrt.a suddenly started exporting them. This broke compatibility
with XP in some cases.

Signed-off-by: Martin Storsjö <martin@martin.st>