headers: Always map e.g. time to _time32 or _time64 for ucrtbase
ucrtbase doesn't have any function named plain 'time'. Force these
to be remapped to the inline form even if inlining otherwise wouldn't
be done. Make sure that we don't keep any non-inline declaration,
because otherwise a caller that tries to take the function address
would end up referring to that.

Signed-off-by: Martin Storsjö <martin@martin.st>
Signed-off-by: Jacek Caban <jacek@codeweavers.com>