msvcrt: Remove "DATA" attributes for _onexit for i386
There is no local wrapper function named _onexit in libmsvcrt.a
or libmingwex.a, and the declaration in headers don't include
any dllimport declaration.

This makes i386 match the other architectures, and fixes linking
to this function with lld. (GNU ld manages to resolve calls to
_onexit into __imp__onexit when there is no other version of the
function available anywhere though.)

The DATA declaration in msvcrt.def.in can be traced back to the
original add of that file in 2008.

Signed-off-by: Martin Storsjö <martin@martin.st>