crt: Allow sleeping for over a second with usleep
Even though the POSIX spec of usleep says that the argument shall
be less than one million. Most unixes allow it and sleeps for over a
second if requested (in practice, at least on Linux/glibc and macOS).

Prefer matching other actual implementations (which ignore this
aspect of the POSIX spec) instead of following the spec strictly.

Signed-off-by: Martin Storsjö <martin@martin.st>