crt: Don't mark the advapi32 MD5Update function as dllimport-only (DATA)
The DATA flag was added without explanation in 7cf3c45e, but
only for lib64, not for lib32.

Marking only this function but not the other MD5 functions was
inconsistent, and could cause linking conflicts with static
libraries that also contained an MD5 implementation with similar
names, if advapi32 was linked before the other implementation.

Signed-off-by: Martin Storsjö <martin@martin.st>