widl: Allow overriding the default include path
At least clang looks for headers in prefix/include in addition to
prefix/triplet/include. This allows configuring widl to look for them
in that directory as well.

When shipping sysroots for more than one or two target architectures,
allowing sharing the headers in a triplet-less include directory saves
a significant amount of space.

Signed-off-by: Martin Storsjö <martin@martin.st>