headers: Change __MSVCRT_VERSION__ to use proper hexadecimal values for CRTs >= 10
This tries to maintain backwards compatibility of enabling UCRT if
either 0xE00 or 0x1400 is specified, but will break user code that
check the value of __MSVCRT_VERSION__ >= 0x1400.

This tries to correct the historical mistake in how the higher values of
this constant is defined, before it goes on further.

This matches how e.g. _WIN32_WINNT_WIN10 is defined.

Signed-off-by: Martin Storsjö <martin@martin.st>