math: Don't use the __INFL constant directly in hypotl
On ARM/ARM64, long double is 64 bit. The __INFL constant is defined
as a bit pattern which only works for the x86 80 bit long floats.

Instead return INFINITY, which the compiler can convert properly to
whichever length long double happens to have.

This was the only function where __INFL is referenced directly.

Signed-off-by: Martin Storsjö <martin@martin.st>