crt: Make the ucrt _vsnwprintf function use UCRTBASE_PRINTF_DEFAULT_WIDE
This function was moved from inline in header to a statically linked
one in 52a34d52d6f5f1789d329abd4e, where the use of this define
was dropped.

As this function is statically linked, the caller can't control
and change the behaviour of the function by defining
UCRTBASE_PRINTF_DEFAULT_WIDE in the calling translation unit,
but we should at least use the headers' default value (which currently
is zero).

Signed-off-by: Martin Storsjö <martin@martin.st>