crt: Move fabsf and nexttowardf to libmsvcr*.a for all archs, and to libucrt*.a for x86
fabsf is available in UCRT on arm32/arm64, but not on x86.

nexttowardf is available in UCRT on all architectures, but this
function takes two parameters, and the second parameter is a long
double in all forms of the function, so we can't use any form
of this function from UCRT on x86.

Signed-off-by: Martin Storsjö <martin@martin.st>