headers: Fix the return value for swprintf and vswprintf for UCRT
According to standard C99, the swprintf and vswprintf functions
return a negative value if the buffer is insufficient (contrary to
snprintf). As far as I can read the standard, the contents of the
buffer in this case is unspecified, but this patch makes it match
MSVC, which does null terminate it.

Signed-off-by: Martin Storsjö <martin@martin.st>