crt: remove winstorecompat duplicated functions from api-ms-win-security-cryptoapi
The API's are forbidden anyway [1] so should be linked through external
libraries. winstorecompat/windowsappcompat has proper replacements.

When used with mincore the only way to link these calls was through
winstorecompat so builds won't be affected by this change.

Technically CryptReleaseContext() is allowed but without other matching API's
as far as I can see. And we cannot use the system one with the custom context
created in winstorecompat.

[1] https://docs.microsoft.com/en-us/uwp/win32-and-com/win32-apis#apis-from-api-ms-win-security-cryptoapi-l1-1-0dll

Signed-off-by: Martin Storsjö <martin@martin.st>