dwrite_1.h: Fix build in C mode
Some classes† in the file declare overloaded functions, and this causes
a C compiler to fail on module including dwrite_1.h.  This commit fixes
that.

To be consistent on changes I rewrote declaration of all classes/structs
in the file, even if one doesn't use overloaded functions.

No changes for compilation using C++ mode.

† IDWriteFont1, IDWriteFontFace1 and IDWriteFactory1.

Signed-off-by: ‡Ruslan Garipov <ruslanngaripov@gmail.com>
Signed-off-by: Liu Hao <lh_mouse@126.com>