crt: Use intermediate libraries to build object files shared across all libmsvcr*.a and libucrt*.a
This avoids compiling the object files that go into libmsvcrt*.a
10 times, and object files that go into libucrt*.a twice.

This reduces the number of built object files for a build for i686
from 921 to 747, and shrinks the generated Makefile.in from 10.0 MB
to 8.9 MB, by emitting per-object rules for fewer object files.

This in itself isn't a huge saving, but there's a general trend of
wanting to move more object files from libmingwex.a into libmvscr*.a,
to allow UCRT builds to use the UCRT provided version of functions
instead of statically linking our own version. And moving object files
in that direction is a bit costly currently as each of them end up
compiled multiple times.

Signed-off-by: Martin Storsjö <martin@martin.st>