headers/stdio: Don't attempt to perform buffer overflow checks when inlining is impossible
We have this in 'gdb/opcodes/z80-dis.c':

   info->fprintf_func = (fprintf_ftype) &sprintf;

This attempts to take the address of `sprintf()` and assign it to a variable,
making it unsuitable for inlining.

With `-D_FORTIFY_SOURCE=2` we could get errors like:

   C:/MSYS2/mingw32/include/stdio.h:389:10: error: invalid use of '__builtin_va_arg_pack ()'
   return __mingw_sprintf( __stream, __format, __builtin_va_arg_pack() );
          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Therefore, it is not safe to make these functions `static inline`. They have
to be `extern inline __attribute__((__gnu_inline__))` instead, to prevent the
inline body from being instantiated where inappropriate. The macro
`__mingw_bos_ovr` is no longer necessary and can be removed.

Revised-by: Martin Storsjö <martin@martin.st>
Signed-off-by: LIU Hao <lh_mouse@126.com>