crt: Implement __gets_chk
When a subset of ssp functions were implemented in
7478c20848f60e6238547db63d1a5bc450d468cc, only functions that were
referenced with _FORTIFY_SOURCE were added. This set of functions
accidentally missed __gets_chk, which we do reference if fortification
is enabled.

Thus, this fixes an accidental omission.

While this function does implicitly reference stderr and the fgets
function, the fgets function symbol is provided by all CRT alternatives.
stderr expands to a call to __acrt_iob_func, and that is provided
by all CRT alternatives, either by the DLL itself or as statically
linked glue code. Therefore, this implementation should be safe
to use with both msvcr*.dll and UCRT.

Signed-off-by: Martin Storsjö <martin@martin.st>