headers: enable some Registry API calls in Win10 19H1 UWP builds
The documentation doesn't say it's allowed but they are allowed by the
Windows Application Certification Kit and the 18362 Windows SDK.

It is not allowed in older SDK. It won't compile or won't link.
The target DLL api-ms-win-core-registry-l1-1-0 will likely not have the
function, so it should not be used when targeting older Windows 10 versions
in UWP mode.

We now have api-ms-win-core-registry-l1-1-0 in windowsapp.

Signed-off-by: LIU Hao <lh_mouse@126.com>