crt: consistently hide winstorecompat wincrypt API's
The x86 bits version hides them, but not the common version. It should
always be hidden as in 9464ea241865f218cdfee9784bb6dc1731a23647.

They are now part of WindowsApp.lib but only enabled for desktop builds.
Desktop users using libwindowsapp.a or libmincore.a won't be
able to use the actual system calls.

Signed-off-by: LIU Hao <lh_mouse@126.com>