headers: allow more wincrypt API's in Win10 RS4 UWP builds
The API's are allowed in windowsapp since RS4 and are allowed by the WACK.
Only the MS header don't specify it properly for WINAPI_FAMILY_PC_APP
but since the DLL is on all WINAPI_FAMILY_DESKTOP_APP and allowed by the
WACK this always works.

CMS_DH_KEY_INFO is needed by CryptSetKeyParam().

Signed-off-by: LIU Hao <lh_mouse@126.com>