crt: remove libgamemode from windowsapp
The 3 API's exported are found in api-ms-win-gaming-expandedresources-l1-1-0
which is part of windowsapp. This is the proper location accessing these
API's from windowsapp [1]. That document says it was introduced in 16299/RS3
but it's actually found there in the 15063/RS2 SDK.

Technically windowsapp also exports/redirects to a lot of non api-ms-***
DLLs, but not this one.

[1] https://learn.microsoft.com/en-us/uwp/win32-and-com/win32-apis#apis-from-api-ms-win-gaming-expandedresources-l1-1-0dll

Signed-off-by: LIU Hao <lh_mouse@126.com>