headers: Add the D3D12 shader compiler interface dxcapi.h
This header is appropriately licensed with a compatible open source
license, the University of Illinois Open Source License (a BSD-like
license).

This header originates from Microsoft, available at e.g.
https://github.com/microsoft/DirectXShaderCompiler/blob/main/include/dxc/dxcapi.h.
That version is incompatible with mingw though, but the Mesa project
has a version that adds mingw compatibility, at
https://gitlab.freedesktop.org/mesa/mesa/-/blob/main/src/microsoft/compiler/dxcapi.h.

This includes the current version of the latter header.

Signed-off-by: Martin Storsjö <martin@martin.st>