headers: Adjust __cpuidex availability for Clang
Clang 18.x was meant to implement __cpuidex (a version was merged
during the development of Clang 17.x, but it was reverted and backed
out before Clang 17.x was completed) - however it was never merged
before Clang 18.x was branched after all. For further history of
the earlier changes within mingw-w64, see
0605217f5d60257a53ec875e2e4ff36a3e14f825 and
2b6c9247613aa830374e3686e09d3b8d582a92a6.

Postpone this change to Clang 19 for now.

This fixes building software that uses __cpuidex with Clang 18.

Hopefully it can either be merged in Clang well in advance before
Clang 19 gets branched, or this issue revisited again long before
that.

Signed-off-by: Martin Storsjö <martin@martin.st>