tools: Remove hardcoded use of -Werror
By hardcoding the use of -Werror, we risk that the build can break
on any newer (or older) toolchain which can give warnings for
unexpected issues.

Fixing such issues is of course desired, but we shouldn't be breaking
the build universally, unless the user explicitly has asked for
this behaviour.

Signed-off-by: Martin Storsjö <martin@martin.st>