crt: Fix comments in msvcrt-common.def.in
The ADD_UNDERSCORE macro adds an alias, without a leading underscore,
for an existing export with a leading underscore.

For some functions, we're not providing these aliases, e.g. because
those symbols are defined by other means, e.g. in source files (such
as CRT_fp10.c and CRT_fp8.c).

This patch improves the style of those cases, which are listed as
commented out instances of the ADD_UNDERSCORE macro.

Signed-off-by: Martin Storsjö <martin@martin.st>