crt: Provide fcloseall as alias to _fcloseall
MSVC does provide this function without an underscore prefix (via
the oldnames library) just like a number of other functions.

In mingw-w64 headers, we already provide a declaration of this
function, just like many others, but it was missing from the
import libraries.

Provide it as an alias, just like we do for many other similar
functions that only exist with a leading underscore in the runtime
DLLs.

Signed-off-by: Martin Storsjö <martin@martin.st>