crt: msvcrt.def.in: Do not export symbol wcsnlen on I386 and X64
Symbol wcsnlen was first introduced in Windows Vista version of msvcrt.dll
and mingw-w64 already provides replacement implementation in libmingwex. So
always use the replacement implementation on I386 and X64 and do not export
it. Also remove DATA keyword and use it in the same way as other wc*
symbols like wcsrtombs.

Signed-off-by: Martin Storsjö <martin@martin.st>