crt: msvcrt.def.in: Reorder symbols chronologically and add comments
Library msvcrt.dll was used by Visual C++ versions 4.2 - 6.0 and is also
part of the Windows system starting with NT 4.0 and 98.

Currently the file msvcrt.def.in contains mix of all symbols in semi
alphabetical order, with ad-hoc symbol aliases which some are not exported
by any msvcrt.dll version (but needed for mingw-w64).

Cleanup symbol definitions, reorder them into chronological order according
to released msvcrt.dll versions and add exhausting comments explaining from
which msvcrt.dll version is symbol present and also if it is being replaced
by other symbol at linking time for backward version compatibility reasons.

If there is a comment "replaced by" at a symbol line, it means that the
symbol is present since particular msvcrt.dll version, but mingw-w64 is
unconditionally replacing it by its own implementation.

If there is a comment "provided by" at a symbol line, it means that the
symbol for some architecture is not present in any version of msvcrt.dll
and mingw-w64 is providing own implementation.

In future this change allows to easily figure out if specific operating
system or Visual C++ version has support for particular function symbol
and if mingw-w64 is going to use that symbol or is replacing it by its own
implementation.

Chronological order of msvcrt.dll versions in which were changes is:

 - Visual C++ 4.2 / Windows NT 4.0
 - Visual C++ 5.0 SP1 / Windows 98 First Edition
 - Visual C++ 6.0 / Windows 98 Second Edition
 - Windows 2000 / Windows ME
 - Windows 2000 SP4 (*)
 - Windows XP
 - Windows XP SP3
 - Windows Server 2003
 - Windows Server 2003 SP1 / Windows XP x64
 - Windows Vista
 - Windows 8
 - Windows 8.1

(*) there is a small exception, new Windows 2000 SP4 symbols are not in
Windows XP until SP3.

msvcrt.def.in currently does not define symbols which were introduced in
Windows 10 (or new) version of msvcrt.dll. There is a TODO comment for it.
Also there is a TODO comment for future cleanup/documentation of ARM
symbols, as this change targets only I386 and X64 versions of msvcrt.dll.

This change as is does not introduce or remove any existing symbol in
msvcrt.def.in. It can be verified by comparing output of the following
command. Replace $PLAT with one of: DEF_I386 DEF_X64 DEF_ARM32 DEF_ARM64

  cpp -x c mingw-w64-crt/lib-common/msvcrt.def.in -Wp,-w -undef -P -Imingw-w64-crt/def-include -D$PLAT | sed -E 's/\s*;.*//' | LC_ALL=C sort -u

Signed-off-by: Martin Storsjö <martin@martin.st>