headers: Don't redefine the __mingw_ovr define in swprintf.inl and wchar.h
This redefining of the macro was added in
824ceb1d1265e257d8b85d8e191f4db2a1e7735c, without an explanation
of why that was done.

If we really do need to use a different inline declaration for
these functions, we should use a different macro, so we don't
alter the meaning of the __mingw_ovr define after including
swprintf.inl or wchar.h.

This practically has the effect, that these inlines are declared
as regular "inline" instead of "static __inline__" when built
in C++ mode with a GCC compatible compiler. This matches how
the __mingw_ovr macro is defined and used for many other inline
functions.

Signed-off-by: Martin Storsjö <martin@martin.st>