headers: Use __mingw_ovr instead of explicitly static inline
When building the C++23 std modules, at least with libc++, the
C++ module needs to reexport a number of standard C functions.

The regular libc++ headers (the ones used even if not using
C++ modules) do this, essentially:

    namespace std {
        using ::ctime;
    }

Thus reexporting the regular C function ctime within the std
namespace.

When building libc++ as a module, this function gets emitted as
part of the C++ std module, like this:

    export namespace std {
        using std::ctime;
    }

This tries to export the function as part of the C++ module. In the
case of our inline functions, this errors out if the inline functions
are static inline:

    <prefix>/share/libc++/v1/std/ctime.inc:20:14: error: using declaration referring to 'ctime' with internal linkage cannot be exported
       20 |   using std::ctime;
          |              ^
    <prefix>/x86_64-w64-mingw32/include/time.h:267:29: note: target of using declaration
      267 | static __inline char *__CRTDECL ctime(const time_t *_Time) { return _ctime64(_Time); }
          |                                 ^

Therefore, prefer using the __mingw_ovr macro for these inline
declarations. This macro expands to regular plain (non-static)
inline in C++ mode, while it still expands to static inline in C mode.

Signed-off-by: Martin Storsjö <martin@martin.st>