crt: Fix utime symbol alias for msvcr80-120 versions
msvcrt versions 8.0 - 12.0 and UCRT do not have _utime symbol.

For these CRT versions (which do not have FIXED_SIZE_SYMBOLS) is later in
the file symbol _utime defined as an alias to _utime32 or _utime64 symbol.

So for these affected CRT versions define utime as an alias to _utime32 or
_utime64.

Signed-off-by: Martin Storsjö <martin@martin.st>