crt: Add guards in func.def.in file
Add once-guard to ensure that the file is processed only once.

Add arch-guard to ensure that DEF_<ARCH> is always defined to prevent an
issue that symbols defined via F32() macro completely disappeared and which
was fixed in previous change.

This change applied before the previous change cause compile errors.

Signed-off-by: Martin Storsjö <martin@martin.st>