crt: msvcrt.def.in: Do not use emulation of Vista+ symbols on ARM32/64
ARM32 and ARM64 always contain all Vista symbols. So use them directly
instead of the mingw-w64 emulation.

The only Vista+ symbol which stay disabled in the def file is vsnprintf.
It is because this Vista+ vsnprintf symbol is not C99 compatible and is
just an alias to _vsnprintf symbol which does not fill nul term byte.

Signed-off-by: Martin Storsjö <martin@martin.st>