crt: Provide __badioinfo variable import symbol for all CRT libraries
Global variable __badioinfo is exported from VC42+ CRT import libraries.
It is not exported from UCRT.

So it is not available in UCRT, msvcrt40.dll and older CRT versions.

Symbol is just a global ioinfo structure with bad (-1) file descriptor.

Provide this structure also in import libraries for older CRT versions and
UCRT, so the symbols will be in all CRT import libraries.

Signed-off-by: Martin Storsjö <martin@martin.st>