crt: msvcr120_app: Do not include aliases for symbols which are not present
msvcr120_app.dll compared to msvcr120.dll is missing not only _getpid symbol
but it is missing also other symbols for which are specified underscore
aliases in crt-aliases.def.in file.

Change NO_GETPID_ALIAS macro to CRTAPP and conditionally disable all
symbols for CRTAPP in crt-aliases.def.in file which are not present in
msvcr120_app.dll library.

Signed-off-by: Martin Storsjö <martin@martin.st>