winpthreads: Fix constructor priority to make it run first
Here is a patch which fixes bugs #988 and #992 following the commit
7b3379.

This commit introduced a new function winpthreads_init in charge of
setting up the _pthread_get_system_time_best_as_file_time and _pthread_get_tick_count_64
pointers using a runtime detection.

This function is defined with the constructor attribute. The problem is
that every static object making use of a constructor is also tagged with
this constructor attribute and the call ordering is unspecified.

For example, when linking code making use of C++11 chrono objects the
libstdc++ code is called before winpthreads_init but makes of the
clock_gettime function which causes a crash.

A minimal reproduction code can be found here in bug #992.

This bug can be fixed by setting a priority of 0 to the
winpthreads_constructor. It is then called before the static
constructors.

Signed-off-by: LIU Hao <lh_mouse@126.com>