headers: Avoid false-positive -Wconversion warnings in isnan() macro
Test case:

  #include <math.h>
  int isnan_wrapper(double value) { return isnan(value); }

Output:

  $ g++ -Wconversion -c snippet.cpp
  warning: conversion to 'float' from 'double' may alter its value
  warning: conversion to 'int' from 'double' may alter its value

First warning is caused by the fact in C++ mode __mingw_choose_expr is
using ternary operator and warning is triggered also by the case which is
not executed.

Second warning is caused by the fact that type of expression
(__builtin_trap(),x) for isnan(x) is double, but return value of
isnan_wrapper() is int. But this expression never returns.

Fix the first problem to use explicit cast for __isnan* calls which
correspondents to matched __mingw_types_compatible_p types.

And fix second problem by changing expression to something which has type
of int.

Bug: https://sourceforge.net/p/mingw-w64/bugs/481/
Signed-off-by: LIU Hao <lh_mouse@126.com>