crt: Move vscanf-family functions from libmingwex.a to individual CRT import libraries
msvcr110 and earlier did not have any v*scanf functions. Because of
that, the mingw-w64-crt contained wrappers that implement the
v*scanf functions based on the non-variadic *scanf functions, by
reading the va_list object and re-passing them as regular parameters
to the *scanf functions.

The sequence involved for vsscanf is:
- stdio/vsscanf2.S, providing vsscanf, calling __ms_vsscanf
- stdio/vsscanf.c, providing __ms_vsscanf, calling __argtos
- stdio/scanf.S, providing __argtos, which converts from a
  va_list back to regular parameters, and calls sscanf()

These functions were provided in libmingwex, making them available
and used with all CRT versions.

However, since msvcr120 (and in UCRT), the proper v*scanf functions
are available. Therefore, move this set of wrappers into the
individual CRT import libraries, allowing using the proper v*scanf
functions on msvcr120 and UCRT.

Signed-off-by: Martin Storsjö <martin@martin.st>